Implement basic Secrets Manager use
Secrets Manager is able to store JSON data natively, values should be
parsed into Microsoft Extensions Configuration Key value pairs.

In order to work the list of secrets on the target account should be
retrieved. A deny on the IAM permissions to this call should be treated
as no secrets stored in Secrets Manager.

The user should be able to change the endpoint and select the profile
that the AWS client should use for authentication.