Improve credentials handling
Adds extra options for controlling the interaction with AWS.

BuildExceptionHandler: Allows consumers to be notified when an exception
occurs while accessing secrets manager. Some other providers such as the
Json file provider have a flag that makes optional the provider. In this
case the provider is optional by default. If an exception is detected
the consumer can rethrow the exception so that the provider does not act
as optional.

AwsCredentialsProfilePath: Controls the path of the AWS credentials
profile path. As per AWS documentation the recomended way of retrieving
credentials is by using credentials profiles, this exposes the
flexibility to specify a profile path other than the default one.
It also helps while testing in order to simulate profiles.

AwsRegionName: The region where the secrets manager will be used. This
must follow AWS region names.