Fix AWSSDK dependent unit tests
Some tests pass due to the fact that the AWSSDK detects a profile or an AWS configuration on the development machine.
For runtime scenarios this is going to be true as long as the code is running on a EC2 instance or a Lambda.

On-premises environments this requirements will be met either by using the explicit options in the configuration.